Relocate ResponseEnvelope to moneybin.protocol
ResponseEnvelope is the shared response shape across MCP, CLI --output
json, and the future HTTP surface. Moving it out of moneybin.mcp makes
the cross-transport contract explicit and prevents weird coupling when
non-MCP transports need it.

Pure rename; no behavior change.
