+++ src/moneybin/mcp/tools/categorize.py
@@ -30,8 +30,8 @@ import duckdb
 
 from moneybin.database import get_database
 from moneybin.mcp.decorator import mcp_tool
-from moneybin.mcp.envelope import ResponseEnvelope, build_envelope
 from moneybin.mcp.namespaces import NamespaceRegistry, ToolDefinition
+from moneybin.protocol.envelope import ResponseEnvelope, build_envelope
 from moneybin.services.auto_rule_service import AutoRuleService
 from moneybin.services.categorization_service import (
     BulkCategorizationResult,
