+++ src/moneybin/mcp/tools/sql.py
@@ -11,9 +11,9 @@ import logging
 
 from moneybin.database import get_database
 from moneybin.mcp.decorator import mcp_tool
-from moneybin.mcp.envelope import ResponseEnvelope, build_envelope
 from moneybin.mcp.namespaces import NamespaceRegistry, ToolDefinition
 from moneybin.mcp.privacy import get_max_rows, validate_read_only_query
+from moneybin.protocol.envelope import ResponseEnvelope, build_envelope
 
 logger = logging.getLogger(__name__)
