+++ src/moneybin/mcp/tools/tax.py
@@ -11,8 +11,8 @@ import logging
 
 from moneybin.database import get_database
 from moneybin.mcp.decorator import mcp_tool
-from moneybin.mcp.envelope import ResponseEnvelope
 from moneybin.mcp.namespaces import NamespaceRegistry, ToolDefinition
+from moneybin.protocol.envelope import ResponseEnvelope
 from moneybin.services.tax_service import TaxService
 
 logger = logging.getLogger(__name__)
