+++ src/moneybin/mcp/tools/transactions.py
@@ -13,8 +13,8 @@ from decimal import Decimal
 
 from moneybin.database import get_database
 from moneybin.mcp.decorator import mcp_tool
-from moneybin.mcp.envelope import ResponseEnvelope
 from moneybin.mcp.namespaces import NamespaceRegistry, ToolDefinition
+from moneybin.protocol.envelope import ResponseEnvelope
 from moneybin.services.transaction_service import TransactionService
 
 logger = logging.getLogger(__name__)
