+++ src/moneybin/protocol/__init__.py
@@ -1,0 +1,1 @@
+"""Cross-transport protocol types shared across MCP, CLI, and future HTTP."""
