+++ src/moneybin/services/account_service.py
@@ -13,7 +13,7 @@ from decimal import Decimal
 from typing import Any
 
 from moneybin.database import Database
-from moneybin.mcp.envelope import ResponseEnvelope, build_envelope
+from moneybin.protocol.envelope import ResponseEnvelope, build_envelope
 from moneybin.tables import DIM_ACCOUNTS, OFX_BALANCES
 
 logger = logging.getLogger(__name__)
