+++ src/moneybin/services/auto_rule_service.py
@@ -20,7 +20,7 @@ import duckdb
 
 from moneybin.config import get_settings
 from moneybin.database import Database
-from moneybin.mcp.envelope import ResponseEnvelope, build_envelope
+from moneybin.protocol.envelope import ResponseEnvelope, build_envelope
 from moneybin.services._text import normalize_description
 from moneybin.services.categorization_service import (
     CategorizationService,
