+++ src/moneybin/services/transaction_service.py
@@ -13,7 +13,7 @@ from decimal import Decimal
 from typing import Any
 
 from moneybin.database import Database
-from moneybin.mcp.envelope import ResponseEnvelope, build_envelope
+from moneybin.protocol.envelope import ResponseEnvelope, build_envelope
 from moneybin.tables import FCT_TRANSACTIONS, TRANSACTION_CATEGORIES
 
 logger = logging.getLogger(__name__)
